# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Dubhu Designer API
  description: API specification for Dubhu Designer services.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
  /tools:
    post:
      summary: Create a new tool
      description: Creates a new tool specification with inputs and outputs.
      operationId: createTool
      tags:
        - Tools
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
            example:
              name: "http_request"
              description: "Sends an HTTP request and returns the response"
              inputs:
                - name: "url"
                  description: "Target URL"
                  required: true
                - name: "method"
                  description: "HTTP method (GET, POST, etc.)"
                  required: false
              outputs:
                - name: "status_code"
                  description: "HTTP status code"
                - name: "body"
                  description: "Response body payload"
      responses:
        '201':
          description: Tool created successfully
          headers:
            ETag:
              description: Strong ETag for the created tool representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tool'
        '400':
          description: Bad request (invalid payload or missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation failed"
                details:
                  - field: "name"
                    reason: "tool name cannot exceed 200 characters"
                    value: "..."
        '409':
          description: Conflict (tool name already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                details:
                  - field: "name"
                    reason: "tool name already exists"
                    value: "http_request"
    get:
      summary: List all tools
      description: Retrieves all registered tool specifications without sorting, filtering, or pagination.
      operationId: listTools
      tags:
        - Tools
      responses:
        '200':
          description: List of tools retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tool'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to query tools"
  /tools/{toolId}:
    get:
      summary: Get tool by ID
      description: Retrieves a specific tool specification by its MongoDB ObjectID.
      operationId: getToolById
      tags:
        - Tools
      parameters:
        - name: toolId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the tool
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
      responses:
        '200':
          description: Tool retrieved successfully
          headers:
            ETag:
              description: Strong ETag for the current tool representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tool'
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Tool not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to find tool: connection refused"
    put:
      summary: Update tool by ID
      description: Updates an existing tool specification. Requires the strong ETag returned when the tool was last read.
      operationId: updateTool
      tags:
        - Tools
      parameters:
        - name: toolId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the tool
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
        - name: If-Match
          in: header
          required: true
          description: Strong ETag from the latest tool representation
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolRequest'
            example:
              name: "http_request_v2"
              description: "Updated HTTP request tool with timeout"
              inputs:
                - name: "url"
                  description: "Target URL"
                  required: true
                - name: "timeout"
                  description: "Request timeout in seconds"
                  required: false
              outputs:
                - name: "status_code"
                  description: "HTTP response status code"
      responses:
        '200':
          description: Tool updated successfully
          headers:
            ETag:
              description: Strong ETag for the updated tool representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tool'
        '400':
          description: Bad request (invalid payload or validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation failed"
                details:
                  - field: "name"
                    reason: "tool name cannot exceed 200 characters"
                    value: "..."
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Tool not found"
        '409':
          description: Conflict (tool name already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          description: The supplied If-Match ETag does not match the current tool representation
        '428':
          description: If-Match header is required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to update tool"
...
components:
  schemas:
    ToolInput:
      type: object
      required:
        - name
        - description
        - required
      properties:
        name:
          type: string
          maxLength: 200
          description: Name of the input parameter
          example: "url"
        description:
          type: string
          maxLength: 2000
          description: Description of the input parameter
          example: "Target URL to send request"
        required:
          type: boolean
          description: Whether the input is required
          example: true
    ToolOutput:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          maxLength: 200
          description: Name of the output parameter
          example: "status_code"
        description:
          type: string
          maxLength: 2000
          description: Description of the output parameter
          example: "HTTP status code"
    CreateToolRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          maxLength: 200
          description: Tool name
          example: "http_request"
        description:
          type: string
          maxLength: 2000
          description: Tool description
          example: "Sends an HTTP request and returns the response"
        inputs:
          type: array
          description: List of tool inputs. Can be omitted, null, or empty array.
          default: []
          items:
            $ref: '#/components/schemas/ToolInput'
        outputs:
          type: array
          description: List of tool outputs. Can be omitted, null, or empty array.
          default: []
          items:
            $ref: '#/components/schemas/ToolOutput'
    UpdateToolRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          maxLength: 200
          description: Tool name
          example: "http_request_v2"
        description:
          type: string
          maxLength: 2000
          description: Tool description
          example: "Updated HTTP request tool with timeout"
        inputs:
          type: array
          description: List of tool inputs. Can be omitted, null, or empty array.
          default: []
          items:
            $ref: '#/components/schemas/ToolInput'
        outputs:
          type: array
          description: List of tool outputs. Can be omitted, null, or empty array.
          default: []
          items:
            $ref: '#/components/schemas/ToolOutput'
    Tool:
      type: object
      required:
        - id
        - name
        - description
        - inputs
        - outputs
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Hexadecimal MongoDB ObjectID
          example: "651da3b57f8641a27e36c579"
        name:
          type: string
          example: "http_request"
        description:
          type: string
          example: "Sends an HTTP request and returns the response"
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/ToolInput'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/ToolOutput'
        createdAt:
          type: string
          format: date-time
          example: "2026-10-03T03:50:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-10-03T03:50:00Z"
...
    FieldError:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
          description: Name or path of the field that failed validation
          example: "name"
        reason:
          type: string
          description: Reason why the field failed validation
          example: "tool name cannot exceed 200 characters"
        value:
          nullable: true
          description: Value that caused the validation failure
          example: "A very long tool name..."
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/FieldError'
            - type: string
              example: "Key: 'CreateToolDTO.Name' Error:Field validation for 'Name' failed on the 'required' tag"